pKVM: VMX: Fix register leak and injection for MSR exits
Like port-IO exits, MSR read/write exits (RDMSR/WRMSR) only use 32-bit
values in RAX and RDX (passed as EDX:EAX), and 32-bit values in RCX
(MSR index).

SDM about RDMSR: "Reads the contents of a 64-bit model specific register
(MSR) specified in the ECX register into registers EDX:EAX.  (On
processors that support the Intel 64 architecture, the high-order 32
bits of RCX are ignored.)"

SDM about WRMSR: "Writes the contents of registers EDX:EAX into the
64-bit model specific register (MSR) specified in the ECX register.  (On
processors that support the Intel 64 architecture, the high-order 32
bits of RCX are ignored.) The contents of the EDX register are copied to
high-order 32 bits of the selected MSR and the contents of the EAX
register are copied to low-order 32 bits of the MSR. (On processors that
support the Intel 64 architecture, the high-order 32 bits of each of RAX
and RDX are ignored.)"

However, the hypervisor was copying the full 64-bit registers:
1. Information Leak: Upper 32 bits of RAX, RDX, and RCX were leaked to
   the host on exit.
2. Register Injection: Host could inject arbitrary data into the
   guest's upper 32 bits of RAX and RDX on RDMSR exit, violating
   architectural zero-extension rules.

Fix this by:
1. On exit (share_protected_vcpu_state): Mask RAX, RDX, and RCX to
   32 bits before sharing them with the host.
2. On entry (update_protected_vcpu_state): Mask the host-supplied RAX
   and RDX values to 32 bits before writing them to the guest,
   enforcing zero-extension.

This was not actually pointed by b/517581231 but it is similar data leak.

Fixes: cdf36a4a528d ("pKVM: VMX: Sync vCPU state for EXIT_REASON_MSR_READ")
Fixes: af531fd1eff6 ("pKVM: VMX: Sync vCPU state for EXIT_REASON_MSR_WRITE")

+++ arch/x86/kvm/vmx/vmx.c
@@ -9805,8 +9805,8 @@ static void update_protected_vcpu_state(struct kvm_vcpu *vcpu,
 
 		if (vmx_get_exit_reason(vcpu).basic == EXIT_REASON_MSR_READ &&
 		    !to_pkvm_vcpu(vcpu)->host_emulated_msr_err) {
-			kvm_rax_write(vcpu, shared_vcpu->arch.regs[VCPU_REGS_RAX]);
-			kvm_rdx_write(vcpu, shared_vcpu->arch.regs[VCPU_REGS_RDX]);
+			kvm_rax_write(vcpu, (u32)shared_vcpu->arch.regs[VCPU_REGS_RAX]);
+			kvm_rdx_write(vcpu, (u32)shared_vcpu->arch.regs[VCPU_REGS_RDX]);
 		}
 
 		WARN_ON_ONCE(kvm_complete_insn_gp(vcpu,
@@ -9944,11 +9944,11 @@ static void share_protected_vcpu_state(struct kvm_vcpu *vcpu,
 		break;
 	}
 	case EXIT_REASON_MSR_WRITE:
-		shared_vcpu->arch.regs[VCPU_REGS_RAX] = kvm_rax_read(vcpu);
-		shared_vcpu->arch.regs[VCPU_REGS_RDX] = kvm_rdx_read(vcpu);
+		shared_vcpu->arch.regs[VCPU_REGS_RAX] = (u32)kvm_rax_read(vcpu);
+		shared_vcpu->arch.regs[VCPU_REGS_RDX] = (u32)kvm_rdx_read(vcpu);
 		fallthrough;
 	case EXIT_REASON_MSR_READ:
-		shared_vcpu->arch.regs[VCPU_REGS_RCX] = kvm_rcx_read(vcpu);
+		shared_vcpu->arch.regs[VCPU_REGS_RCX] = (u32)kvm_rcx_read(vcpu);
 		break;
 	case EXIT_REASON_EPT_VIOLATION:
 		to_vmx(shared_vcpu)->exit_gpa = vmcs_read64(GUEST_PHYSICAL_ADDRESS);
